renamed noise model files; changed noise model def in tests
