+++ probeye/definition/__init__.py
@@ -1,6 +1,6 @@
 # module imports
 from probeye.definition import inference_problem
 from probeye.definition import forward_model
-from probeye.definition import noise_model
+from probeye.definition import likelihood_model
 from probeye.definition import parameter
 from probeye.definition import prior
