+++ tests/integration_tests/test_maximum_likelihood.py
@@ -17,7 +17,7 @@ import matplotlib.pyplot as plt
 from probeye.definition.inference_problem import InferenceProblem
 from probeye.definition.forward_model import ForwardModelBase
 from probeye.definition.sensor import Sensor
-from probeye.definition.noise_model import NormalNoiseModel
+from probeye.definition.likelihood_model import NormalNoiseModel
 
 # local imports (inference related)
 from probeye.inference.scipy_.solver import ScipySolver
@@ -92,11 +92,6 @@ class TestProblem(unittest.TestCase):
         linear_model = LinearModel(["m", "b"], [isensor], [osensor])
         problem.add_forward_model("LinearModel", linear_model)
 
-        # add the noise model to the problem
-        problem.add_noise_model(
-            NormalNoiseModel(prms_def={"sigma": "std"}, sensors=osensor)
-        )
-
         # ============================================================================ #
         #                    Add test data to the Inference Problem                    #
         # ============================================================================ #
@@ -116,9 +111,6 @@ class TestProblem(unittest.TestCase):
             sensor_values={isensor.name: x_test, osensor.name: y_test},
         )
 
-        # give problem overview
-        problem.info()
-
         # plot the true and noisy data
         if plot:
             plt.scatter(x_test, y_test, label="measured data", s=10, c="red", zorder=10)
@@ -128,6 +120,18 @@ class TestProblem(unittest.TestCase):
             plt.legend()
             plt.tight_layout()
             plt.draw()  # does not stop execution
+
+        # ============================================================================ #
+        #                              Add noise model(s)                              #
+        # ============================================================================ #
+
+        # add the noise model to the problem
+        problem.add_noise_model(
+            NormalNoiseModel(prms_def={"sigma": "std"}, sensors=osensor)
+        )
+
+        # give problem overview
+        problem.info()
 
         # ============================================================================ #
         #                    Solve problem with inference engine(s)                    #
