+++ tests/integration_tests/test_spatial_correlation.py
@@ -23,7 +23,7 @@ import matplotlib.pyplot as plt
 from probeye.definition.inference_problem import InferenceProblem
 from probeye.definition.forward_model import ForwardModelBase
 from probeye.definition.sensor import Sensor
-from probeye.definition.noise_model import NormalNoiseModel
+from probeye.definition.likelihood_model import NormalNoiseModel
 
 # local imports (testing related)
 from tests.integration_tests.subroutines import run_inference_engines
@@ -190,15 +190,6 @@ class TestProblem(unittest.TestCase):
         linear_model = LinearModel(["a", "b"], [], [osensor])
         problem.add_forward_model("LinearModel", linear_model)
 
-        # add the noise model to the problem
-        noise_model = NormalNoiseModel(
-            sensors=osensor,
-            corr="x",
-            corr_model="exp",
-            prms_def=[{"sigma": "std"}, "l_corr"],
-        )
-        problem.add_noise_model(noise_model)
-
         # ============================================================================ #
         #                    Add test data to the Inference Problem                    #
         # ============================================================================ #
@@ -233,6 +224,22 @@ class TestProblem(unittest.TestCase):
             plt.legend()
             plt.tight_layout()
             plt.draw()  # plt.draw() does not stop execution
+
+        # ============================================================================ #
+        #                              Add noise model(s)                              #
+        # ============================================================================ #
+
+        # add the noise model to the problem
+        noise_model = NormalNoiseModel(
+            sensors=osensor,
+            corr="x",
+            corr_model="exp",
+            prms_def=[{"sigma": "std"}, "l_corr"],
+        )
+        problem.add_noise_model(noise_model)
+
+        # give problem overview
+        problem.info()
 
         # ============================================================================ #
         #                    Solve problem with inference engine(s)                    #
