+++ tests/unit_tests/definition/test_inference_problem.py
@@ -10,7 +10,7 @@ import numpy as np
 from probeye.definition.inference_problem import InferenceProblem
 from probeye.definition.forward_model import ForwardModelBase
 from probeye.definition.sensor import Sensor
-from probeye.definition.noise_model import NoiseModelBase
+from probeye.definition.likelihood_model import NoiseModelBase
 
 
 class TestProblem(unittest.TestCase):
